requirements.txt
datasets==3.2.0
fastwer==0.1.3
matplotlib==3.5.1
nltk==3.9.1
numpy==1.21.5
pandas==2.2.3
PyYAML==5.4.1
PyYAML==6.0.2
Requests==2.32.3
sacrebleu==2.5.1
scipy==1.8.0
spacy_stanza==1.0.4
stanza==1.10.1
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0.dev0
